+++ docs/docs/tracing.md
@@ -8,17 +8,11 @@ sidebar_position: 5
 
 Ознакомиться с использованием **OpenTelemetry** можно на [официальном сайте](https://opentelemetry.io/).
 
-Пакет **isomorph** на данный момент предоставляет несколько простых утилит для базовой трассировки ответов express-приложения.
+Пакет **isomorph** на данный момент предоставляет несколько простых утилит для базовой трассировки.
 
-## tracingMiddleware (express)
+## Axios
 
-**tracingMiddleware** возвращает готовый промежуточный слой для express-приложения. Этот промежуточный слой будет создавать span на каждый запрос и завершать его при ответе.
-
-Контекст и корневой span доступны на `res.locals.tracing.rootContext` и `res.locals.tracing.rootSpan` соответственно.
-
-## tracingMiddleware (middleware-axios)
-
-**tracingMiddleware** возвращает готовый промежуточный слой для `AxiosInstanceWrapper`. Этот промежуточный слой будет создавать span на каждый запрос и завершать его при ответе.
+Пакет предоставляет функцию **tracingMiddleware** которая возвращает промежуточный слой для `AxiosInstanceWrapper`. Этот промежуточный слой будет создавать span на каждый запрос и завершать его при ответе.
 
 ```ts
 import express from 'express';
